CI: Calculate Nightly versions even after downgrading version string
Don't naively use the latest Nightly release (and assume it was either
the previous nightly for this exact major.minor.patch,
or we are starting over at nightly0).

Instead, look for the latest Nightly release that matches our
major.minor.patch and iterate on _that_, defaulting to nightly0
only if such a matching release was not found in the first page of
API results.